Add CLang code for RDtsc, combine wrapper headers.
